refactor: extract MCP tool handlers from route.ts into separate modules
Split the 1625-line route.ts into focused handler modules:
- handlers/shared.ts: types, Zod schemas, constants, helpers
- handlers/search.ts: search_icons tool
- handlers/icons.ts: get_icon, get_multiple_icons tools
- handlers/starter-packs.ts: get_starter_pack tool
- handlers/resources.ts: MCP resource registrations
- handlers/bundles.ts: authenticated bundle tools (list/get/create)
- handlers/index.ts: barrel exports

route.ts now contains only MCP protocol handling (234 lines).
All tools and resources are functionally identical.
