+++ src/app/api/mcp/handlers/index.ts
@@ -1,0 +1,6 @@
+export { registerSearchTools } from "./search";
+export { registerIconTools } from "./icons";
+export { registerStarterPackTools } from "./starter-packs";
+export { registerResources } from "./resources";
+export { registerBundleTools } from "./bundles";
+export type { AuthContext } from "./shared";
