+++ src/app/api/mcp/route.ts
@@ -11,33 +11,17 @@
 
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { WebStandardStreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/webStandardStreamableHttp.js";
-import { z } from "zod";
-import {
-  searchIcons,
-  getSearchCount,
-  getIconById,
-  getIconsByIds,
-  getIconsByNames,
-  getSources,
-  getCategories,
-  getTotalIconCount,
-} from "@/lib/queries";
-import {
-  convertIconToFormat,
-  generateReactBundle,
-  generateSvgBundle,
-  generateJsonBundle,
-} from "@/lib/icon-converters";
-import { normalizeIcons, normalizeIcon, STANDARD_VIEWBOX } from "@/lib/icon-utils";
-import { STARTER_PACKS } from "@/lib/starter-packs";
 import { logger } from "@/lib/logger";
-import { createAdminClient } from "@/lib/supabase/admin";
 import { validateApiToken } from "@/lib/auth/api-token";
 import { checkPublicRateLimit, getRateLimitHeaders } from "@/lib/rate-limit";
-
-// Constants
-const CHARACTER_LIMIT = 100000; // Maximum response size in characters
-const VALID_PACK_IDS = STARTER_PACKS.map((p) => p.id);
+import {
+  type AuthContext,
+  registerSearchTools,
+  registerIconTools,
+  registerStarterPackTools,
+  registerResources,
+  registerBundleTools,
+} from "./handlers";
 
 // CORS headers for cross-origin access (v0, Bolt, Lovable, etc.)
 const CORS_HEADERS = {
@@ -46,88 +30,6 @@ const CORS_HEADERS = {
   "Access-Control-Allow-Headers":
     "Content-Type, Accept, Authorization, Mcp-Session-Id, Mcp-Protocol-Version, Last-Event-Id",
 };
-
-// Reusable Zod schemas
-const iconIdSchema = z
-  .string()
-  .min(3)
-  .max(100)
-  .regex(
-    /^[a-z][a-z0-9-]*:[a-z0-9][a-z0-9-]*$/i,
-    "Invalid iconId format. Expected 'source:name' (e.g., 'lucide:arrow-right')"
-  )
-  .describe("Icon ID in format 'source:name' (e.g., 'lucide:arrow-right')");
-
-const formatSchema = z
-  .enum(["svg", "react", "vue", "svelte", "json"])
-  .default("react")
-  .describe("Output format");
-
-const sizeSchema = z
-  .number()
-  .int()
-  .min(8)
-  .max(512)
-  .default(24)
-  .describe("Icon size in pixels");
-
-const strokeWidthSchema = z
-  .number()
-  .min(0.5)
-  .max(4)
-  .default(2)
-  .describe("Stroke width");
-
-const normalizeStrokesSchema = z
-  .boolean()
-  .default(false)
-  .describe("Normalize stroke widths across icons (skips fill-based icons automatically)");
-
-const outputSchema = z
-  .enum(["individual", "bundle"])
-  .default("bundle")
-  .describe("Output mode: 'bundle' (single file, 70% smaller) or 'individual' (separate components)");
-
-const sourceSchema = z
-  .enum([
-    "lucide",
-    "phosphor",
-    "hugeicons",
-    "heroicons",
-    "tabler",
-    "feather",
-    "remix",
-    "simple-icons",
-    "iconoir",
-  ])
-  .optional()
-  .describe("Filter by icon library");
-
-const bundleNameSchema = z.string().min(1).max(80).describe("Bundle name");
-
-const bundleDescriptionSchema = z
-  .string()
-  .max(300)
-  .optional()
-  .nullable()
-  .describe("Bundle description");
-
-const bundleIconIdsSchema = z
-  .array(iconIdSchema)
-  .min(1)
-  .max(200)
-  .describe("List of icon IDs to save in the bundle");
-
-const strokePresetSchema = z
-  .enum(["thin", "regular", "bold"])
-  .optional()
-  .describe("Stroke weight preset");
-
-const viewBoxSchema = z
-  .string()
-  .max(50)
-  .default(STANDARD_VIEWBOX)
-  .describe("Target viewBox when normalizing");
 
 // Helper to add CORS headers to a response
 function withCors(response: Response): Response {
@@ -142,64 +44,10 @@ function withCors(response: Response): Response {
   });
 }
 
-// Helper to truncate response if too large
-function truncateIfNeeded(text: string, limit: number = CHARACTER_LIMIT): string {
-  if (text.length <= limit) return text;
-  return (
-    text.slice(0, limit) +
-    `\n\n[Response truncated. Original size: ${text.length} characters. Use pagination or filters to reduce results.]`
-  );
-}
-
-// Helper to get comment syntax based on format
-function getCommentSyntax(format: string): { start: string; end: string } {
-  switch (format) {
-    case "svg":
-      return { start: "<!-- ", end: " -->" };
-    case "react":
-    case "json":
-      return { start: "// ", end: "" };
-    case "vue":
-      return { start: "<!-- ", end: " -->" };
-    case "svelte":
-      return { start: "<!-- ", end: " -->" };
-    default:
-      return { start: "// ", end: "" };
-  }
-}
-
-// Helper to format batch icons with identifying comments
-function formatBatchIconsText(
-  icons: Array<{ id?: string; name: string; code: string; error?: string }>,
-  format: string
-): string {
-  const { start, end } = getCommentSyntax(format);
-  const separator = "\n" + "=".repeat(60) + "\n";
-
-  return icons
-    .map((icon, index) => {
-      const iconId = icon.id || icon.name;
-      const header = `${start}[${index + 1}] ${iconId}${icon.error ? " (ERROR)" : ""}${end}`;
-
-      if (icon.error) {
-        return `${header}\n${start}Error: ${icon.error}${end}`;
-      }
-
-      return `${header}\n${icon.code}`;
-    })
-    .join(separator);
-}
-
-// Auth context for authenticated tools
-interface AuthContext {
-  userId: string;
-  isPro: boolean;
-}
-
 // Create MCP server with all tools and resources
 function createMcpServer(authContext?: AuthContext) {
   const isAuthenticated = !!authContext?.userId;
-  
+
   const server = new McpServer({
     name: "unicon",
     version: "1.1.0",
@@ -224,1255 +72,15 @@ QUICK START:
 Read unicon://instructions for detailed usage.`,
   });
 
-  // ============================================
-  // TOOL: search_icons
-  // ============================================
-  server.registerTool(
-    "search_icons",
-    {
-      title: "Search Icons",
-      description: `Search icons across 9 libraries. Set includeCode=true to get ready-to-use components in one call.
-
-Args:
-  - query (string): Search query (e.g., 'arrow', 'dashboard', 'user profile')
-  - includeCode (boolean, optional): Return icon code with results (default: false)
-  - format (string, optional): Code format when includeCode=true (default: react)
-  - normalizeStrokes (boolean, optional): Normalize stroke widths, skipping fill icons (default: false)
-  - source (string, optional): Filter by library
-  - limit (number, optional): Max results (1-100, default: 20)
-
-Examples:
-  - Search only: query="arrow"
-  - Search + code: query="arrow", includeCode=true, limit=5
-  - Normalized: query="arrow", includeCode=true, normalizeStrokes=true, strokeWidth=2`,
-      inputSchema: z
-        .object({
-          query: z.string().min(1).max(200).describe("Search query"),
-          includeCode: z
-            .boolean()
-            .default(false)
-            .describe("Include icon source code in results (returns bundle format)"),
-          format: formatSchema,
-          strokeWidth: strokeWidthSchema,
-          normalizeStrokes: normalizeStrokesSchema,
-          source: sourceSchema,
-          category: z.string().max(50).optional().describe("Filter by category"),
-          limit: z
-            .number()
-            .int()
-            .min(1)
-            .max(100)
-            .default(20)
-            .describe("Maximum results to return"),
-          offset: z
-            .number()
-            .int()
-            .min(0)
-            .default(0)
-            .describe("Number of results to skip for pagination"),
-        })
-        .strict(),
-      outputSchema: z
-        .object({
-          query: z.string(),
-          total: z.number(),
-          offset: z.number(),
-          limit: z.number(),
-          has_more: z.boolean(),
-          results: z.array(
-            z.object({
-              id: z.string(),
-              name: z.string(),
-              normalizedName: z.string(),
-              source: z.string(),
-              category: z.string().nullable(),
-              tags: z.array(z.string()),
-            })
-          ),
-        })
-        .strict(),
-      annotations: {
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-        openWorldHint: false,
-      },
-    },
-    async (params) => {
-      const limit = params.limit ?? 20;
-      const offset = params.offset ?? 0;
-      const includeCode = params.includeCode ?? false;
-      const format = params.format as "svg" | "react" | "vue" | "svelte" | "json";
-      const normalizeStrokes = params.normalizeStrokes ?? false;
-      const strokeWidth = params.strokeWidth ?? 2;
-
-      // Build search params with proper typing for exactOptionalPropertyTypes
-      const filterParams: {
-        query: string;
-        sourceId?: string;
-        category?: string;
-      } = {
-        query: params.query,
-      };
-
-      if (params.source !== undefined) {
-        filterParams.sourceId = params.source;
-      }
-      if (params.category !== undefined) {
-        filterParams.category = params.category;
-      }
-
-      // Run results query and count query in parallel for accurate pagination
-      const [dbResults, totalCount] = await Promise.all([
-        searchIcons({
-          ...filterParams,
-          limit,
-          offset,
-        }),
-        getSearchCount(filterParams),
-      ]);
-
-      // Calculate has_more from accurate total
-      const hasMore = offset + dbResults.length < totalCount;
-
-      // If includeCode is true, return bundled code with results
-      if (includeCode && dbResults.length > 0) {
-        // Validate format supports bundling
-        if (format !== "react" && format !== "svg" && format !== "json") {
-          return {
-            content: [
-              {
-                type: "text",
-                text: `Error: includeCode only supports format="react", "svg", or "json". Got "${format}". Use output="individual" for Vue/Svelte.`,
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        // Apply stroke normalization if requested (skips fill-based icons)
-        const iconsToBundle = normalizeStrokes
-          ? normalizeIcons(dbResults, { strokeWidth, skipFillIcons: true })
-          : dbResults;
-
-        let bundleText: string;
-        if (format === "react") {
-          bundleText = generateReactBundle(iconsToBundle, { strokeWidth });
-        } else if (format === "svg") {
-          bundleText = generateSvgBundle(iconsToBundle, { strokeWidth });
-        } else {
-          bundleText = generateJsonBundle(iconsToBundle);
-        }
-
-        const output = {
-          query: params.query,
-          total: totalCount,
-          offset,
-          limit,
-          has_more: hasMore,
-          format,
-          results: dbResults.map((icon) => ({
-            id: icon.id,
-            name: icon.name,
-            normalizedName: icon.normalizedName,
-            source: icon.sourceId,
-          })),
-          code: bundleText,
-        };
-
-        const text = truncateIfNeeded(bundleText);
-
-        return {
-          content: [{ type: "text", text }],
-          structuredContent: output,
-        };
-      }
-
-      // Standard search results (no code)
-      const output = {
-        query: params.query,
-        total: totalCount,
-        offset,
-        limit,
-        has_more: hasMore,
-        results: dbResults.map((icon) => ({
-          id: icon.id,
-          name: icon.name,
-          normalizedName: icon.normalizedName,
-          source: icon.sourceId,
-          category: icon.category,
-          tags: icon.tags,
-        })),
-      };
-
-      const text = truncateIfNeeded(JSON.stringify(output, null, 2));
-
-      return {
-        content: [{ type: "text", text }],
-        structuredContent: output,
-      };
-    }
-  );
-
-  // ============================================
-  // TOOL: get_icon
-  // ============================================
-  server.registerTool(
-    "get_icon",
-    {
-      title: "Get Icon",
-      description: `Retrieve source code for a specific icon in various formats.
-
-Args:
-  - iconId (string): Icon ID in format 'source:name' (e.g., 'lucide:arrow-right')
-  - format (string, optional): Output format - svg, react, vue, svelte, json (default: react)
-  - size (number, optional): Icon size in pixels (default: 24)
-  - strokeWidth (number, optional): Stroke width for line icons (default: 2)
-  - normalizeStrokes (boolean, optional): Normalize stroke widths, skipping fill icons (default: false)
-
-Returns:
-  Object with iconId, format, and code.
-
-Examples:
-  - "get the lucide arrow-right icon" -> iconId="lucide:arrow-right"
-  - "get phosphor house icon as SVG" -> iconId="phosphor:house", format="svg"`,
-      inputSchema: z
-        .object({
-          iconId: iconIdSchema,
-          format: formatSchema,
-          size: sizeSchema,
-          strokeWidth: strokeWidthSchema,
-          normalizeStrokes: normalizeStrokesSchema,
-        })
-        .strict(),
-      outputSchema: z
-        .object({
-          iconId: z.string(),
-          format: z.string(),
-          code: z.string(),
-        })
-        .strict(),
-      annotations: {
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-        openWorldHint: false,
-      },
-    },
-    async (params) => {
-      const icon = await getIconById(params.iconId);
-      if (!icon) {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Error: Icon not found: "${params.iconId}". Use search_icons to find valid icon IDs, or browse https://unicon.sh`,
-            },
-          ],
-          isError: true,
-        };
-      }
-
-      const format = params.format as "svg" | "react" | "vue" | "svelte" | "json";
-      const normalizeStrokes = params.normalizeStrokes ?? false;
-      const strokeWidth = params.strokeWidth ?? 2;
-
-      // Apply stroke normalization if requested (skips fill-based icons)
-      const iconToConvert = normalizeStrokes
-        ? normalizeIcon(icon, { strokeWidth, skipFillIcons: true })
-        : icon;
-
-      const code = await convertIconToFormat(iconToConvert, format, {
-        size: params.size,
-        strokeWidth,
-      });
-
-      const output = {
-        iconId: params.iconId,
-        format,
-        code,
-      };
-
-      return {
-        content: [{ type: "text", text: code }],
-        structuredContent: output,
-      };
-    }
-  );
-
-  // ============================================
-  // TOOL: get_multiple_icons
-  // ============================================
-  server.registerTool(
-    "get_multiple_icons",
-    {
-      title: "Get Multiple Icons",
-      description: `Retrieve multiple icons at once (max 50 per request).
-
-Args:
-  - iconIds (array): Array of icon IDs (e.g., ['lucide:arrow-right', 'lucide:home'])
-  - output (string, optional): 'bundle' (default, single file) or 'individual' (separate components)
-  - format (string, optional): svg, react, vue, svelte, json (default: react)
-  - size (number, optional): Icon size in pixels (default: 24)
-  - strokeWidth (number, optional): Stroke width (default: 2)
-  - normalizeStrokes (boolean, optional): Normalize stroke widths, skipping fill icons (default: false)
-
-Returns:
-  Single copy-pasteable file with all icons (bundle) or individual components.`,
-      inputSchema: z
-        .object({
-          iconIds: z
-            .array(iconIdSchema)
-            .min(1)
-            .max(50)
-            .describe("Array of icon IDs to retrieve"),
-          output: outputSchema,
-          format: formatSchema,
-          size: sizeSchema,
-          strokeWidth: strokeWidthSchema,
-          normalizeStrokes: normalizeStrokesSchema,
-        })
-        .strict(),
-      outputSchema: z
-        .object({
-          format: z.string(),
-          total: z.number(),
-          successful: z.number(),
-          icons: z.array(
-            z.object({
-              id: z.string(),
-              name: z.string(),
-              code: z.string(),
-              error: z.string().optional(),
-            })
-          ),
-        })
-        .strict(),
-      annotations: {
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-        openWorldHint: false,
-      },
-    },
-    async (params) => {
-      const format = params.format as "svg" | "react" | "vue" | "svelte" | "json";
-      const outputMode = params.output ?? "bundle";
-      const normalizeStrokes = params.normalizeStrokes ?? false;
-      const strokeWidth = params.strokeWidth ?? 2;
-
-      // Batch fetch all icons in a single query (eliminates N+1)
-      const fetchedIcons = await getIconsByIds(params.iconIds);
-      const iconsById = new Map(fetchedIcons.map((icon) => [icon.id, icon]));
-
-      // Get ordered list of found icons (preserving request order)
-      const orderedIcons = params.iconIds
-        .map((id) => iconsById.get(id))
-        .filter((icon): icon is NonNullable<typeof icon> => icon !== undefined);
-
-      // Bundle mode: single file output (70% smaller)
-      if (outputMode === "bundle") {
-        // Validate format supports bundling
-        if (format !== "react" && format !== "svg" && format !== "json") {
-          return {
-            content: [
-              {
-                type: "text",
-                text: `Error: Bundle mode only supports format="react", "svg", or "json". Got "${format}". Use output="individual" for Vue/Svelte.`,
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        // Apply stroke normalization if requested (skips fill-based icons)
-        const iconsToBundle = normalizeStrokes
-          ? normalizeIcons(orderedIcons, { strokeWidth, skipFillIcons: true })
-          : orderedIcons;
-
-        let bundleText: string;
-        if (format === "react") {
-          bundleText = generateReactBundle(iconsToBundle, {
-            size: params.size,
-            strokeWidth,
-          });
-        } else if (format === "svg") {
-          bundleText = generateSvgBundle(iconsToBundle, {
-            size: params.size,
-            strokeWidth,
-          });
-        } else {
-          bundleText = generateJsonBundle(iconsToBundle);
-        }
-
-        const text = truncateIfNeeded(bundleText);
-
-        return {
-          content: [{ type: "text", text }],
-          structuredContent: {
-            format,
-            output: "bundle",
-            total: params.iconIds.length,
-            successful: orderedIcons.length,
-            code: bundleText,
-          },
-        };
-      }
-
-      // Individual mode: separate components (original behavior)
-      const results = await Promise.all(
-        params.iconIds.map(async (iconId) => {
-          try {
-            const icon = iconsById.get(iconId);
-            if (!icon) {
-              return {
-                id: iconId,
-                name: "",
-                code: "",
-                error: "Icon not found. Use search_icons to find valid IDs.",
-              };
-            }
-
-            // Apply stroke normalization if requested (skips fill-based icons)
-            const iconToConvert = normalizeStrokes
-              ? normalizeIcon(icon, { strokeWidth, skipFillIcons: true })
-              : icon;
-
-            const code = await convertIconToFormat(iconToConvert, format, {
-              size: params.size,
-              strokeWidth,
-            });
-
-            return { id: icon.id, name: icon.name, code };
-          } catch (error) {
-            return {
-              id: iconId,
-              name: "",
-              code: "",
-              error: error instanceof Error ? error.message : "Unknown error",
-            };
-          }
-        })
-      );
-
-      const output = {
-        format,
-        output: "individual",
-        total: results.length,
-        successful: results.filter((r) => !r.error).length,
-        icons: results,
-      };
-
-      // Format text with identifying comments for each icon
-      const formattedText = formatBatchIconsText(results, format);
-      const text = truncateIfNeeded(formattedText);
-
-      return {
-        content: [{ type: "text", text }],
-        structuredContent: output,
-      };
-    }
-  );
-
-  // ============================================
-  // TOOL: get_starter_pack
-  // ============================================
-  server.registerTool(
-    "get_starter_pack",
-    {
-      title: "Get Starter Pack",
-      description: `Get a curated starter pack of icons for common use cases.
-
-Popular packs: shadcn-ui, dashboard, ecommerce, navigation, developer, brand-ai
-
-Args:
-  - packId (string): Starter pack ID
-  - output (string, optional): 'bundle' (default, single file) or 'individual' (separate components)
-  - format (string, optional): svg, react, vue, svelte, json (default: react)
-  - size (number, optional): Icon size (default: 24)
-  - strokeWidth (number, optional): Stroke width (default: 2)
-  - normalizeStrokes (boolean, optional): Normalize stroke widths, skipping fill icons (default: false)
-
-Returns:
-  Single copy-pasteable file with all icons (bundle) or individual components.`,
-      inputSchema: z
-        .object({
-          packId: z
-            .string()
-            .min(1)
-            .max(50)
-            .describe(
-              `Starter pack ID. Available: ${VALID_PACK_IDS.slice(0, 6).join(", ")}... Use unicon://starter_packs for full list.`
-            ),
-          output: outputSchema,
-          format: formatSchema,
-          size: sizeSchema,
-          strokeWidth: strokeWidthSchema,
-          normalizeStrokes: normalizeStrokesSchema,
-        })
-        .strict(),
-      outputSchema: z
-        .object({
-          packId: z.string(),
-          packName: z.string(),
-          description: z.string(),
-          format: z.string(),
-          totalIcons: z.number(),
-          retrievedIcons: z.number(),
-          icons: z.array(
-            z.object({
-              name: z.string(),
-              code: z.string(),
-              error: z.string().optional(),
-            })
-          ),
-        })
-        .strict(),
-      annotations: {
-        readOnlyHint: true,
-        destructiveHint: false,
-        idempotentHint: true,
-        openWorldHint: false,
-      },
-    },
-    async (params) => {
-      const pack = STARTER_PACKS.find((p) => p.id === params.packId);
-      if (!pack) {
-        return {
-          content: [
-            {
-              type: "text",
-              text: `Error: Starter pack not found: "${params.packId}". Available packs: ${VALID_PACK_IDS.join(", ")}. Use the unicon://starter_packs resource to see details.`,
-            },
-          ],
-          isError: true,
-        };
-      }
-
-      const format = params.format as "svg" | "react" | "vue" | "svelte" | "json";
-      const outputMode = params.output ?? "bundle";
-      const normalizeStrokes = params.normalizeStrokes ?? false;
-      const strokeWidth = params.strokeWidth ?? 2;
-
-      // Batch fetch all icons (eliminates N+1)
-      const fetchedIcons = await getIconsByNames(pack.iconNames);
-      const iconsByName = new Map(
-        fetchedIcons.map((icon) => [icon.normalizedName.toLowerCase(), icon])
-      );
-
-      // Get ordered list of found icons (preserving pack order)
-      const orderedIcons = pack.iconNames
-        .map((name) => iconsByName.get(name.toLowerCase()))
-        .filter((icon): icon is NonNullable<typeof icon> => icon !== undefined);
-
-      // Bundle mode: single file output (70% smaller)
-      if (outputMode === "bundle") {
-        // Validate format supports bundling
-        if (format !== "react" && format !== "svg" && format !== "json") {
-          return {
-            content: [
-              {
-                type: "text",
-                text: `Error: Bundle mode only supports format="react", "svg", or "json". Got "${format}". Use output="individual" for Vue/Svelte.`,
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        // Apply stroke normalization if requested (skips fill-based icons)
-        const iconsToBundle = normalizeStrokes
-          ? normalizeIcons(orderedIcons, { strokeWidth, skipFillIcons: true })
-          : orderedIcons;
-
-        let bundleText: string;
-        if (format === "react") {
-          bundleText = generateReactBundle(iconsToBundle, {
-            size: params.size,
-            strokeWidth,
-          });
-        } else if (format === "svg") {
-          bundleText = generateSvgBundle(iconsToBundle, {
-            size: params.size,
-            strokeWidth,
-          });
-        } else {
-          bundleText = generateJsonBundle(iconsToBundle);
-        }
-
-        const text = truncateIfNeeded(bundleText);
-
-        return {
-          content: [{ type: "text", text }],
-          structuredContent: {
-            packId: pack.id,
-            packName: pack.name,
-            description: pack.description,
-            format,
-            output: "bundle",
-            totalIcons: pack.iconNames.length,
-            retrievedIcons: orderedIcons.length,
-            code: bundleText,
-          },
-        };
-      }
-
-      // Individual mode: separate components (original behavior)
-      const results = await Promise.all(
-        pack.iconNames.map(async (iconName) => {
-          try {
-            const icon = iconsByName.get(iconName.toLowerCase());
-            if (!icon) {
-              return { name: iconName, code: "", error: "Icon not found in database" };
-            }
-
-            // Apply stroke normalization if requested (skips fill-based icons)
-            const iconToConvert = normalizeStrokes
-              ? normalizeIcon(icon, { strokeWidth, skipFillIcons: true })
-              : icon;
-
-            const code = await convertIconToFormat(iconToConvert, format, {
-              size: params.size,
-              strokeWidth,
-            });
-
-            return { name: icon.normalizedName, code };
-          } catch (error) {
-            return {
-              name: iconName,
-              code: "",
-              error: error instanceof Error ? error.message : "Unknown error",
-            };
-          }
-        })
-      );
-
-      const output = {
-        packId: pack.id,
-        packName: pack.name,
-        description: pack.description,
-        format,
-        output: "individual",
-        totalIcons: pack.iconNames.length,
-        retrievedIcons: results.filter((r) => !r.error).length,
-        icons: results,
-      };
-
-      // Format text with identifying comments for each icon
-      const header = `${getCommentSyntax(format).start}Starter Pack: ${pack.name} (${pack.id})${getCommentSyntax(format).end}\n${getCommentSyntax(format).start}${pack.description}${getCommentSyntax(format).end}\n\n`;
-      const formattedText = header + formatBatchIconsText(results, format);
-      const text = truncateIfNeeded(formattedText);
-
-      return {
-        content: [{ type: "text", text }],
-        structuredContent: output,
-      };
-    }
-  );
-
-  // ============================================
-  // RESOURCES
-  // ============================================
-
-  server.registerResource(
-    "sources",
-    "unicon://sources",
-    {
-      description: "List all available icon libraries",
-      mimeType: "application/json",
-    },
-    async () => {
-      const sources = await getSources();
-      return {
-        contents: [
-          {
-            uri: "unicon://sources",
-            mimeType: "application/json",
-            text: JSON.stringify(sources, null, 2),
-          },
-        ],
-      };
-    }
-  );
-
-  server.registerResource(
-    "categories",
-    "unicon://categories",
-    {
-      description: "List all icon categories",
-      mimeType: "application/json",
-    },
-    async () => {
-      const categories = await getCategories();
-      return {
-        contents: [
-          {
-            uri: "unicon://categories",
-            mimeType: "application/json",
-            text: JSON.stringify({ categories }, null, 2),
-          },
-        ],
-      };
-    }
-  );
-
-  server.registerResource(
-    "stats",
-    "unicon://stats",
-    {
-      description: "Total icon count and per-library statistics",
-      mimeType: "application/json",
-    },
-    async () => {
-      const totalIcons = await getTotalIconCount();
-      const sources = await getSources();
-      const stats = {
-        totalIcons,
-        sources: sources.map((s) => ({
-          id: s.id,
-          name: s.name,
-          count: s.totalIcons,
-        })),
-      };
-      return {
-        contents: [
-          {
-            uri: "unicon://stats",
-            mimeType: "application/json",
-            text: JSON.stringify(stats, null, 2),
-          },
-        ],
-      };
-    }
-  );
-
-  server.registerResource(
-    "starter_packs",
-    "unicon://starter_packs",
-    {
-      description: "Curated icon packs for common use cases",
-      mimeType: "application/json",
-    },
-    async () => {
-      const packs = {
-        total: STARTER_PACKS.length,
-        packs: STARTER_PACKS.map((pack) => ({
-          id: pack.id,
-          name: pack.name,
-          description: pack.description,
-          color: pack.color,
-          iconCount: pack.iconNames.length,
-          icons: pack.iconNames,
-        })),
-      };
-      return {
-        contents: [
-          {
-            uri: "unicon://starter_packs",
-            mimeType: "application/json",
-            text: JSON.stringify(packs, null, 2),
-          },
-        ],
-      };
-    }
-  );
-
-  // ============================================
-  // INSTRUCTIONS RESOURCE - Tool Discoverability
-  // ============================================
-  server.registerResource(
-    "instructions",
-    "unicon://instructions",
-    {
-      description: "How to use Unicon MCP - read this first to understand available tools",
-      mimeType: "text/markdown",
-    },
-    async () => {
-      const instructions = `# Unicon MCP Server
-
-Access 19,000+ icons from 9 libraries (Lucide, Phosphor, Hugeicons, Heroicons, Tabler, Feather, Remix, Simple Icons, Iconoir).
-
-## Available Tools
-
-### 1. search_icons
-Search for icons by keyword. Returns matching icons with IDs.
-
-**Example:**
-\`\`\`json
-{ "query": "arrow", "limit": 10 }
-\`\`\`
-
-**With code included:**
-\`\`\`json
-{ "query": "arrow", "includeCode": true, "format": "react", "limit": 5 }
-\`\`\`
-
-### 2. get_icon
-Get a single icon's code by ID.
-
-**Example:**
-\`\`\`json
-{ "iconId": "lucide:arrow-right", "format": "react" }
-\`\`\`
-
-Icon IDs use format \`source:name\` (e.g., "lucide:home", "phosphor:user", "heroicons:cog").
-
-### 3. get_multiple_icons
-Get multiple icons at once (up to 50). Returns a single bundled file.
-
-**Example:**
-\`\`\`json
-{ "iconIds": ["lucide:home", "lucide:settings", "lucide:user"], "format": "react" }
-\`\`\`
-
-### 4. get_starter_pack
-Get a curated set of icons for common use cases.
-
-**Popular packs:** shadcn-ui, dashboard, ecommerce, navigation, developer, brand-ai
-
-**Example:**
-\`\`\`json
-{ "packId": "shadcn-ui", "format": "react" }
-\`\`\`
-
-### 5. Authenticated bundle tools (requires UNICON_API_TOKEN)
-Manage your saved bundles (Pro feature).
-
-- list_my_bundles: List your saved bundles
-- get_my_bundle: Fetch a saved bundle by ID (with optional code output)
-- create_my_bundle: Save a new bundle to your account
-
-## Supported Formats
-
-- \`react\` (default) - React component with TypeScript
-- \`svg\` - Raw SVG markup
-- \`vue\` - Vue 3 SFC component
-- \`svelte\` - Svelte component
-- \`json\` - JSON with SVG paths
-
-## Quick Workflows
-
-**Adding icons to a React project:**
-1. Search: \`search_icons({ query: "dashboard", includeCode: true, limit: 10 })\`
-2. Or get specific: \`get_multiple_icons({ iconIds: ["lucide:home", "lucide:settings"] })\`
-3. Copy the returned code to \`src/components/icons/\`
-
-**Starting a new project:**
-1. Get a starter pack: \`get_starter_pack({ packId: "shadcn-ui" })\`
-2. Save to your icons directory
-
-## Icon Libraries
-
-| Library | Prefix | Style |
-|---------|--------|-------|
-| Lucide | lucide: | Outline, clean |
-| Phosphor | phosphor: | Multiple weights |
-| Heroicons | heroicons: | Outline/solid |
-| Tabler | tabler: | Consistent stroke |
-| Hugeicons | hugeicons: | Modern, detailed |
-| Feather | feather: | Minimal |
-| Remix | remix: | Versatile |
-| Simple Icons | simple-icons: | Brand logos |
-| Iconoir | iconoir: | European style |
-`;
-      return {
-        contents: [
-          {
-            uri: "unicon://instructions",
-            mimeType: "text/markdown",
-            text: instructions,
-          },
-        ],
-      };
-    }
-  );
-
-  // ============================================
-  // AUTHENTICATED TOOLS (all logged-in users)
-  // ============================================
+  // Register public tools
+  registerSearchTools(server);
+  registerIconTools(server);
+  registerStarterPackTools(server);
+  registerResources(server);
+
+  // Register authenticated tools
   if (isAuthenticated && authContext) {
-    // TOOL: list_my_bundles
-    server.registerTool(
-      "list_my_bundles",
-      {
-        title: "List My Bundles",
-        description: `List all your saved icon bundles. Requires authentication via UNICON_API_TOKEN.
-
-Returns:
-  Array of bundles with id, name, description, and icon count.
-
-Example:
-  list_my_bundles({})`,
-        inputSchema: z.object({}).strict(),
-        outputSchema: z
-          .object({
-            bundles: z.array(
-              z.object({
-                id: z.string(),
-                name: z.string(),
-                description: z.string().nullable(),
-                icon_count: z.number(),
-                is_public: z.boolean(),
-                share_slug: z.string().nullable(),
-              })
-            ),
-            total: z.number(),
-          })
-          .strict(),
-        annotations: {
-          readOnlyHint: true,
-          destructiveHint: false,
-          idempotentHint: true,
-          openWorldHint: false,
-        },
-      },
-      async () => {
-        const supabase = createAdminClient();
-
-        const { data: bundles, error } = await supabase
-          .from("bundles")
-          .select("id, name, description, icon_count, is_public, share_slug, created_at, updated_at")
-          .eq("user_id", authContext.userId)
-          .order("updated_at", { ascending: false });
-
-        if (error) {
-          return {
-            content: [{ type: "text", text: `Error fetching bundles: ${error.message}` }],
-            isError: true,
-          };
-        }
-
-        const output = {
-          bundles: bundles || [],
-          total: bundles?.length || 0,
-        };
-
-        return {
-          content: [{ type: "text", text: JSON.stringify(output, null, 2) }],
-          structuredContent: output,
-        };
-      }
-    );
-
-    // TOOL: get_my_bundle
-    server.registerTool(
-      "get_my_bundle",
-      {
-        title: "Get My Bundle",
-        description: `Get icons from one of your saved bundles. Requires authentication via UNICON_API_TOKEN.
-
-Args:
-  - bundleId (string): The bundle ID (UUID)
-  - format (string, optional): Output format - svg, react, json (default: react)
-  - includeCode (boolean, optional): Include generated code (default: true)
-
-Returns:
-  Bundle metadata, list of icons, and optionally generated code.
-
-Example:
-  get_my_bundle({ bundleId: "abc-123", format: "react" })`,
-        inputSchema: z
-          .object({
-            bundleId: z.string().uuid().describe("Bundle ID"),
-            format: z
-              .enum(["svg", "react", "json"])
-              .default("react")
-              .describe("Output format"),
-            includeCode: z
-              .boolean()
-              .default(true)
-              .describe("Include generated code in response"),
-          })
-          .strict(),
-        outputSchema: z
-          .object({
-            bundle: z.object({
-              id: z.string(),
-              name: z.string(),
-              description: z.string().nullable(),
-              icon_count: z.number(),
-              icons: z.array(
-                z.object({
-                  id: z.string(),
-                  name: z.string(),
-                  normalizedName: z.string(),
-                  source: z.string(),
-                })
-              ),
-            }),
-            format: z.string(),
-            code: z.string().optional(),
-          })
-          .strict(),
-        annotations: {
-          readOnlyHint: true,
-          destructiveHint: false,
-          idempotentHint: true,
-          openWorldHint: false,
-        },
-      },
-      async (params) => {
-        const supabase = createAdminClient();
-        const format = params.format as "svg" | "react" | "json";
-        const includeCode = params.includeCode ?? true;
-
-        // Fetch the bundle
-        const { data: bundle, error } = await supabase
-          .from("bundles")
-          .select("*")
-          .eq("id", params.bundleId)
-          .eq("user_id", authContext.userId)
-          .single();
-
-        if (error || !bundle) {
-          return {
-            content: [
-              {
-                type: "text",
-                text: `Error: Bundle not found. Use list_my_bundles to see your available bundles.`,
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        // Get icon IDs from bundle
-        const iconIds = (bundle.icons as Array<{ id: string }>)?.map((i) => i.id) || [];
-
-        if (iconIds.length === 0) {
-          const output = {
-            bundle: {
-              id: bundle.id,
-              name: bundle.name,
-              description: bundle.description,
-              icon_count: 0,
-              icons: [],
-            },
-            format,
-            code: includeCode ? "// Empty bundle - no icons" : undefined,
-          };
-          return {
-            content: [{ type: "text", text: "Bundle is empty - no icons saved." }],
-            structuredContent: output,
-          };
-        }
-
-        // Fetch full icon data
-        const icons = await getIconsByIds(iconIds);
-
-        // Apply normalization if configured
-        const normalizedIcons = bundle.normalize_strokes
-          ? normalizeIcons(icons, {
-              strokeWidth: bundle.target_stroke_width || 2,
-              skipFillIcons: true,
-            })
-          : icons;
-
-        // Generate code if requested
-        let code: string | undefined;
-        if (includeCode) {
-          const strokeWidth = bundle.target_stroke_width || 2;
-          switch (format) {
-            case "react":
-              code = generateReactBundle(normalizedIcons, { strokeWidth });
-              break;
-            case "svg":
-              code = generateSvgBundle(normalizedIcons, { strokeWidth });
-              break;
-            case "json":
-              code = generateJsonBundle(normalizedIcons);
-              break;
-          }
-        }
-
-        const output = {
-          bundle: {
-            id: bundle.id,
-            name: bundle.name,
-            description: bundle.description,
-            icon_count: icons.length,
-            icons: icons.map((icon) => ({
-              id: icon.id,
-              name: icon.name,
-              normalizedName: icon.normalizedName,
-              source: icon.sourceId,
-            })),
-          },
-          format,
-          code,
-        };
-
-        const text = includeCode && code ? truncateIfNeeded(code) : JSON.stringify(output, null, 2);
-
-        return {
-          content: [{ type: "text", text }],
-          structuredContent: output,
-        };
-      }
-    );
-
-    // TOOL: create_my_bundle
-    server.registerTool(
-      "create_my_bundle",
-      {
-        title: "Create My Bundle",
-        description: `Create and save a new bundle to your account. Requires authentication via UNICON_API_TOKEN.
-
-Args:
-  - name (string): Bundle name
-  - iconIds (string[]): Array of icon IDs (source:name)
-  - description (string, optional): Bundle description
-  - strokePreset (string, optional): thin | regular | bold
-  - normalizeStrokes (boolean, optional): Normalize stroke widths across icons
-  - targetStrokeWidth (number, optional): Stroke width when normalizeStrokes=true
-  - normalizeViewbox (boolean, optional): Normalize viewBox sizes
-  - targetViewbox (string, optional): Target viewBox when normalizeViewbox=true
-
-Returns:
-  Bundle metadata with id and icon count.
-
-Example:
-  create_my_bundle({ name: "dashboard", iconIds: ["lucide:home", "lucide:settings"] })`,
-        inputSchema: z
-          .object({
-            name: bundleNameSchema,
-            description: bundleDescriptionSchema,
-            iconIds: bundleIconIdsSchema,
-            strokePreset: strokePresetSchema,
-            normalizeStrokes: normalizeStrokesSchema,
-            targetStrokeWidth: strokeWidthSchema.optional(),
-            normalizeViewbox: z.boolean().default(false),
-            targetViewbox: viewBoxSchema,
-          })
-          .strict(),
-        outputSchema: z
-          .object({
-            bundle: z.object({
-              id: z.string(),
-              name: z.string(),
-              description: z.string().nullable(),
-              icon_count: z.number(),
-              is_public: z.boolean().optional(),
-              share_slug: z.string().nullable().optional(),
-            }),
-          })
-          .strict(),
-        annotations: {
-          readOnlyHint: false,
-          destructiveHint: false,
-          idempotentHint: false,
-          openWorldHint: false,
-        },
-      },
-      async (params) => {
-        const supabase = createAdminClient();
-
-        const uniqueIconIds: string[] = [];
-        const seen = new Set<string>();
-        for (const id of params.iconIds) {
-          if (!seen.has(id)) {
-            seen.add(id);
-            uniqueIconIds.push(id);
-          }
-        }
-
-        if (uniqueIconIds.length === 0) {
-          return {
-            content: [{ type: "text", text: "Error: No icon IDs provided." }],
-            isError: true,
-          };
-        }
-
-        const icons = await getIconsByIds(uniqueIconIds);
-        const iconMap = new Map(icons.map((icon) => [icon.id, icon]));
-        const missing = uniqueIconIds.filter((id) => !iconMap.has(id));
-
-        if (missing.length > 0) {
-          return {
-            content: [
-              {
-                type: "text",
-                text: `Error: ${missing.length} icon(s) not found: ${missing.join(", ")}`,
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        const orderedIcons = uniqueIconIds
-          .map((id) => iconMap.get(id))
-          .filter((icon): icon is NonNullable<typeof icon> => !!icon);
-
-        const bundleIcons = orderedIcons.map((icon) => ({
-          id: icon.id,
-          name: icon.name,
-          normalizedName: icon.normalizedName,
-          sourceId: icon.sourceId,
-          svg: icon.content,
-          viewBox: icon.viewBox,
-          strokeWidth: icon.strokeWidth,
-          defaultFill: icon.defaultFill,
-          defaultStroke: icon.defaultStroke,
-        }));
-
-        const { data, error } = await supabase.rpc("create_bundle_atomic", {
-          p_user_id: authContext.userId,
-          p_name: params.name,
-          p_description: params.description ?? null,
-          p_icons: bundleIcons,
-          p_stroke_preset: params.strokePreset ?? null,
-          p_normalize_strokes: params.normalizeStrokes ?? false,
-          p_target_stroke_width: params.targetStrokeWidth ?? null,
-          p_normalize_viewbox: params.normalizeViewbox ?? false,
-          p_target_viewbox: params.targetViewbox ?? STANDARD_VIEWBOX,
-        });
-
-        if (error) {
-          return {
-            content: [{ type: "text", text: `Error creating bundle: ${error.message}` }],
-            isError: true,
-          };
-        }
-
-        const result = data?.[0];
-        if (!result?.success) {
-          return {
-            content: [
-              {
-                type: "text",
-                text: `Error creating bundle: ${result?.error ?? "Unknown error"}`,
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        const bundle = (result.bundle || {}) as {
-          id?: string;
-          name?: string;
-          description?: string | null;
-          icon_count?: number;
-          is_public?: boolean;
-          share_slug?: string | null;
-        };
-
-        const output = {
-          bundle: {
-            id: bundle.id ?? "",
-            name: bundle.name ?? params.name,
-            description: bundle.description ?? null,
-            icon_count: bundle.icon_count ?? bundleIcons.length,
-            is_public: bundle.is_public ?? false,
-            share_slug: bundle.share_slug ?? null,
-          },
-        };
-
-        return {
-          content: [{ type: "text", text: JSON.stringify(output, null, 2) }],
-          structuredContent: output,
-        };
-      }
-    );
+    registerBundleTools(server, authContext);
   }
 
   return server;
@@ -1511,7 +119,7 @@ async function getAuthContext(request: Request): Promise<AuthContext | undefined
 async function handleMcpRequest(request: Request, method: string): Promise<Response> {
   // Extract auth context if token provided
   const authContext = await getAuthContext(request);
-  
+
   const server = createMcpServer(authContext);
   const transport = createTransport();
